avoid warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@11332 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/envisat/EnvisatFile.c
@@ -1039,7 +1039,7 @@ int EnvisatFile_SetKeyValueAsInt( EnvisatFile *self,
         return FAILURE;
     }
 
-    sprintf( format, "%%+0%dd", strlen(prototype_value) );
+    sprintf( format, "%%+0%dd", (int) strlen(prototype_value) );
     sprintf( string_value, format, value );
 
     return EnvisatFile_SetKeyValueAsString( self, mph_or_sph, key, string_value );
